+++ scripts/generate_allreduce_data.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+DTYPES="int float16 float32"
+for SIZE in {2..8}
+do
+    for DTYPE in ${DTYPES}
+    do
+        echo "Allreduce (size: ${SIZE} | dtype: ${DTYPE})"
+        make allreduce -e WORLD_SIZE=${SIZE} HCCL_DTYPE=${DTYPE}
+    done
+done