+++ scripts/plotter.py
@@ -4,10 +4,14 @@
 To generate the allreduce data.
 
 #!/bin/bash
+DTYPES="int float16 float32"
 for SIZE in {2..8}
 do
-   echo "Allreduce (size: ${SIZE})"
-   make allreduce -e WORLD_SIZE=${SIZE}
+    for DTYPE in ${DTYPES}
+    do
+        echo "Allreduce (size: ${SIZE} | dtype: ${DTYPE})"
+        make allreduce -e WORLD_SIZE=${SIZE} HCCL_DTYPE=${DTYPE}
+    done
 done
 """