+++ tests/test_commons.py
@@ -1,6 +1,6 @@
 import torch
 
-from pytorch_hccl_tests.commons import get_device, get_dtype
+from pytorch_hccl_tests.commons import get_device, get_dtype, is_integral
 
 
 def test_get_device():
@@ -33,3 +33,19 @@ def test_get_dtype_double():
 def test_get_dtype_float64():
     dtype = get_dtype("float64")
     assert dtype == torch.float64
+
+
+def test_is_integral_int():
+    assert is_integral("int")
+
+
+def test_is_integral_long():
+    assert is_integral("long")
+
+
+def test_is_integral_float():
+    assert not is_integral("float")
+
+
+def test_is_integral_double():
+    assert not is_integral("double")